MarketLists

Documentation for MarketLists.

MarketLists.getExchangeMembers

MarketLists.getExchangeMembers

MarketLists.getExchangeOptions

MarketLists.getExchangeOptions

MarketLists.getIndexMembersFile

MarketLists.getIndexMembersFile

MarketLists.getIndexOptions

MarketLists.getIndexOptions

MarketLists.parseExchangeResponseToDataFrame

MarketLists.parseExchangeResponseToDataFrame

MarketLists.parseIndexMembersFileToDataFrame

MarketLists.parseIndexMembersFileToDataFrame

MarketLists.getExchangeMembers — Method

getExchangeMembers(exchange::String)::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}

Get all Members of the given Exchange. Supported Exchanges can be seen by calling the getExchangeOptions function. This method calls the nasdaq API at https://api.nasdaq.com/api to get the data.

Arguments

– exchange::String: The name of the Exchange to fetch members for.

Example

Getting Data ...
Dict{Symbol, Union{Int16, String, Vector{UInt8}}} with 3 entries:
  :code    => 200
  :message => "OK"
  :body    => "{"data":{"headers":{"symbol":"Symbol","name":"Name","lastsale":"Last Sale","netchange":
  [...]

MarketLists.getExchangeOptions — Method

getExchangeOptions()::Vector{String}

Get a List of all supported Exchanges. These are the Exchanges for which the getExchangeMembers function will return values.

See getExchangeMembers.

Example

getExchangeOptions()
3-element Vector{String}:
 "NASDAQ"
 "NYSE"
 "AMEX"

MarketLists.getIndexMembersFile — Method

getIndexMembersFile(index::String)::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}

Get all Members of the given Index ETF. Supported ETFs can be seen by calling the getIndexOptions function. This method used the State Street Global Advisors, and Invesco web pages to look up the ETF Membership information.

The results will be saved in an XSLX file for symbols serviced from State Street, and CSV for those from Invesco.

Arguments

– index::String: The name of the Index ETF to fetch members for.

Example

i = getIndexMembersFile("DIA")
Getting holdings for DIA ... 
Dict{Symbol, Union{Int16, String, Vector{UInt8}}} with 3 entries:
  :file    => "/tmp/holdings-daily-us-en-dia.xlsx"
  :code    => 200
  :message => "OK"

MarketLists.getIndexOptions — Method

function getIndexOptions()::Vector{Pair{String, String}}

Get a list of all supported Index ETFs These are the ETF symbols for which the getIndexMembersFile function will return values.

See getIndexMembersFile.

Example

getIndexOptions()
12-element Vector{Pair{String, String}}:
 "CQQQ" => "China Tech ETF"
  "DGT" => "DOW Global"
  "DIA" => "DOW 30"
  "MDY" => "S&P 400"
 "ONEK" => "Russell 1000"
  "QQQ" => "Nasdaq 100"
 "QQQJ" => "Nasdaq Next Gen 100"
  "SLY" => "S&P 600"
  "SMD" => "S&P 1000"
  "SPY" => "S&P 500"
 "THRK" => "Russell 3000"
 "TWOK" => "Russell 2000"

MarketLists.parseExchangeResponseToDataFrame — Method

parseExchangeResponseToDataFrame(httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}, exchange::String)::DataFrame

Parse the JSON result from the getExchangeMembers function into a DataFrame. If the getExchangeMembers function does not return HTTP 200, or the results can not be parsed the DataFrame will contain an error message.

Arguments

– httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}: The return from the getExchangeMembers function. – exchange::String: The name of the Exchange members were fetched for. This is used in determining how to parse the data.

Example

parseExchangeResponseToDataFrame(e, "NYSE")
3012x11 DataFrame
  Row | symbol  name                               lastsale    netchange   pctchange volume   marketCap   country       ipoYear          industry                           sector
      | String  String                             Float64     Float64     Float64   Int64    Union       String        Nothing          String                             String
  -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
    1 | A       Agilent Technologies Inc. Common   140.33      -2.9        -2.025    1088380  4.1518e10   United States                  Electrical Products                Industrials
    2 | AA      Alcoa Corporation Common Stock      50.17      -3.52       -6.556    6297138  8.87854e9                                  Metal Fabrications                 Industrials
    3 | AAC     Ares Acquisition Corporation Cla    10.25      -0.01       -0.097      27539  1.28125e9                                  Blank Checks                       Finance
    4 | AAIC    Arlington Asset Investment Corp      3.04       0.01        0.33       62580  8.62347e7   United States                  Real Estate Investment Trusts      Real Estate
    [...]

MarketLists.parseIndexMembersFileToDataFrame — Method

parseIndexMembersFileToDataFrame(httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}, index::String)::DataFrame

Parse the file downloaded when calling the getIndexMembersFile function.

Arguments

– httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}: The return from the getIndexMembersFile function. – index::String: The name of the Index ETF members were fetched for. This is used in determining how to parse the data.

Example

parseIndexMembersFileToDataFrame(i, "DIA")
31x8 DataFrame
 Row | symbol    name                               Identifier  SEDOL       Weight    Sector                  SharesHeld   Local Currency
     | String    String                             String      String      Float64   String                  Float64      String
 ----------------------------------------------------------------------------------------------------------------------------------------
   1 | UNH       UnitedHealth Group Incorporated    91324P10    2917766     9.47776   Health Care               5.64911e6  USD
   2 | GS        Goldman Sachs Group Inc.           38141G10    2407966     7.03863   Financials                5.64911e6  USD
   3 | HD        Home Depot Inc.                    43707610    2434209     5.8787    Consumer Discretionary    5.64911e6  USD
   4 | MCD       McDonald's Corporation             58013510    2550707     5.33521   Consumer Discretionary    5.64911e6  USD