Indexes

getIndexOptions

MarketLists.getIndexOptions — Function

function getIndexOptions()::Vector{Pair{String, String}}

Get a list of all supported Index ETFs These are the ETF symbols for which the getIndexMembersFile function will return values.

See getIndexMembersFile.

Example

getIndexOptions()
12-element Vector{Pair{String, String}}:
 "CQQQ" => "China Tech ETF"
  "DGT" => "DOW Global"
  "DIA" => "DOW 30"
  "MDY" => "S&P 400"
 "ONEK" => "Russell 1000"
  "QQQ" => "Nasdaq 100"
 "QQQJ" => "Nasdaq Next Gen 100"
  "SLY" => "S&P 600"
  "SMD" => "S&P 1000"
  "SPY" => "S&P 500"
 "THRK" => "Russell 3000"
 "TWOK" => "Russell 2000"

getIndexMembersFile

MarketLists.getIndexMembersFile — Function

getIndexMembersFile(index::String)::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}

Get all Members of the given Index ETF. Supported ETFs can be seen by calling the getIndexOptions function. This method used the State Street Global Advisors, and Invesco web pages to look up the ETF Membership information.

The results will be saved in an XSLX file for symbols serviced from State Street, and CSV for those from Invesco.

Arguments

– index::String: The name of the Index ETF to fetch members for.

Example

i = getIndexMembersFile("DIA")
Getting holdings for DIA ... 
Dict{Symbol, Union{Int16, String, Vector{UInt8}}} with 3 entries:
  :file    => "/tmp/holdings-daily-us-en-dia.xlsx"
  :code    => 200
  :message => "OK"

parseIndexMembersFileToDataFrame

MarketLists.parseIndexMembersFileToDataFrame — Function

parseIndexMembersFileToDataFrame(httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}, index::String)::DataFrame

Parse the file downloaded when calling the getIndexMembersFile function.

Arguments

– httpRet::Dict{Symbol, Union{String, Int16, Vector{UInt8}}}: The return from the getIndexMembersFile function. – index::String: The name of the Index ETF members were fetched for. This is used in determining how to parse the data.

Example

parseIndexMembersFileToDataFrame(i, "DIA")
31x8 DataFrame
 Row | symbol    name                               Identifier  SEDOL       Weight    Sector                  SharesHeld   Local Currency
     | String    String                             String      String      Float64   String                  Float64      String
 ----------------------------------------------------------------------------------------------------------------------------------------
   1 | UNH       UnitedHealth Group Incorporated    91324P10    2917766     9.47776   Health Care               5.64911e6  USD
   2 | GS        Goldman Sachs Group Inc.           38141G10    2407966     7.03863   Financials                5.64911e6  USD
   3 | HD        Home Depot Inc.                    43707610    2434209     5.8787    Consumer Discretionary    5.64911e6  USD
   4 | MCD       McDonald's Corporation             58013510    2550707     5.33521   Consumer Discretionary    5.64911e6  USD